+++ modules/core/dependency/python-ihm/ihm/reader.py
@@ -1447,7 +1447,10 @@ class _AssemblyDetailsHandler(Handler):
             a.append(self.sysr.ranges.get(obj, entity_poly_segment_id))
 
         self.system._make_complete_assembly()
-        tup_complete = tuple(self.system.complete_assembly)
+        # The order of components should not matter, so put in a consistent
+        # order so we can compare against other assemblies
+        complete = sorted(self.system.complete_assembly,
+                          key=lambda x: id(x))
 
         for a in self.system.orphan_assemblies:
             # Any EntityRange or AsymUnitRange which covers an entire entity,
@@ -1455,7 +1458,7 @@ class _AssemblyDetailsHandler(Handler):
             a[:] = [self._handle_component(x) for x in a]
             # If the input file defines the complete assembly, transfer
             # user-provided info to system.complete_assembly
-            if tuple(a) == tup_complete:
+            if sorted(a, key=lambda x: id(x)) == complete:
                 self.system.complete_assembly.name = a.name
                 self.system.complete_assembly.description = a.description